requirements.txt
streamlit>=1.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
psutil>=5.9.0
Pillow>=9.5.0
python-dotenv>=1.0.0